Lesson 7

+++ src/Controller/NotelistController.php
@@ -4,11 +4,14 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
+use App\Entity\Category;
 use App\Entity\Note;
 use App\Repository\NoteRepository;
-use Exception;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
@@ -23,103 +26,24 @@ class NotelistController extends AbstractController
         $this->noteRepository = $noteRepository;
     }
 
-    private array $categories = [
-        1 => [
-            'title' => 'My Summer weekends',
-            'notes' => [1, 2, 3]
-        ],
-        2 => [
-            'title' => 'My favorite books review',
-            'notes' => [4, 5, 6]
-        ],
-        3 => [
-            'title' => 'My friends hobbies',
-            'notes' => [7, 8, 9]
-        ]
-    ];
-
-    private array $notes = [
-        1 => [
-            'id' => 1,
-            'title' => 'Some note 1',
-            'text' => 'Lorem ipsun 1',
-            'category_id' => 1
-        ],
-        2 => [
-            'id' => 2,
-            'title' => 'Some note 2',
-            'text' => 'Lorem ipsun 2',
-            'category_id' => 1
-        ],
-        3 => [
-            'id' => 3,
-            'title' => 'Some note 3',
-            'text' => 'Lorem ipsun 3',
-            'category_id' => 1
-        ],
-        4 => [
-            'id' => 4,
-            'title' => 'Some note 4',
-            'text' => 'Lorem ipsun 4',
-            'category_id' => 2
-        ],
-        5 => [
-            'id' => 5,
-            'title' => 'Some note 5',
-            'text' => 'Lorem ipsun 5',
-            'category_id' => 2
-        ],
-        6 => [
-            'id' => 6,
-            'title' => 'Some note 6',
-            'text' => 'Lorem ipsun 6',
-            'category_id' => 2
-        ],
-        7 => [
-            'id' => 7,
-            'title' => 'Some note 7',
-            'text' => 'Lorem ipsun 7',
-            'category_id' => 3
-        ],
-        8 => [
-            'id' => 8,
-            'title' => 'Some note 8',
-            'text' => 'Lorem ipsun 8',
-            'category_id' => 3
-        ],
-        9 => [
-            'id' => 9,
-            'title' => 'Some note 9',
-            'text' => 'Lorem ipsun 9',
-            'category_id' => 3
-        ]
-    ];
-
     /**
      * @Route(name="list_all")
      */
-    public function listAll(): Response
+    public function listAll(EntityManagerInterface $em): Response
     {
         return $this->render('notelist/list.html.twig', [
-            'notes' => $this->notes,
+            'notes' => $em->getRepository(Note::class)->findAll()
         ]);
     }
 
     /**
      * @Route("/{categoryId}", name="list_by_category", requirements={"categoryId"="\d+"})
      */
-    public function listByCategory(string $categoryId): Response
+    public function listByCategory(string $categoryId, EntityManagerInterface $em): Response
     {
-        if (!isset($this->categories[(int) $categoryId])) {
-            throw new Exception('You ask for category that not exists');
-        }
-
-        $category = $this->categories[(int) $categoryId] ?? null;
-        $notesIds = $category['notes'];
-
-        $notes = array_filter($this->notes, function (array $note) use ($notesIds) {
-           return in_array($note['id'], $notesIds, true);
-        });
+        $notes = $em->getRepository(Note::class)->findBy([
+            'category' => (int) $categoryId
+        ]);
 
         return $this->render('notelist/list.html.twig', [
             'notes' => $notes
@@ -129,23 +53,12 @@ class NotelistController extends AbstractController
     /**
      * @Route("/{categoryId}/{noteId}", name="get", requirements={"categoryId"="\d+", "noteId"="\d+"})
      */
-    public function getAction(string $categoryId, string $noteId): Response
+    public function getAction(string $categoryId, string $noteId, EntityManagerInterface $em): Response
     {
-        if (!isset($this->categories[(int) $categoryId])) {
-            throw new Exception('You ask for category that not exists');
-        }
-
-        $category = $this->categories[(int) $categoryId] ?? null;
-        $notesIds = $category['notes'];
-
-        $notes = array_filter($this->notes, function (array $note) use ($notesIds) {
-            return in_array($note['id'], $notesIds, true);
-        });
-        if (!isset($notes[(int) $noteId])) {
-            throw new Exception('There is no note in selected category');
-        }
-
-        $note = $notes[(int) $noteId];
+        $note = $em->getRepository(Note::class)->findOneBy([
+            'category' => (int) $categoryId,
+            'id' => $noteId
+        ]);
 
         return $this->render('notelist/get.html.twig', [
             'note' => $note
@@ -153,38 +66,52 @@ class NotelistController extends AbstractController
     }
 
     /**
-     * @Route("/create", name="create")
+     * @Route("/create", name="create", methods={"GET", "POST"})
      */
-    public function createAction(): Response
+    public function createAction(Request $request, EntityManagerInterface $em): Response
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $newNote = new Note();
-        $newNote
-            ->setTitle('New note title')
-            ->setText('New note text');
+        if ($request->getMethod() === 'GET') {
+            $categories = $em->getRepository(Category::class)->findAll();
 
-        $entityManager->persist($newNote);
-        $entityManager->flush();
+            return $this->render('notelist/create.html.twig', [
+                'categories' => $categories
+            ]);
+        }
 
-        $notes = $this->noteRepository->findAll();
+        // TODO: add data validation
+        $title = (string) $request->request->get('title');
+        $text = (string) $request->request->get('text');
 
-        return $this->render('notelist/list.html.twig', [
-            'notes' => $notes,
-        ]);
+        $categoryId = (int) $request->request->get('category_id');
+        $category = $em->getRepository(Category::class)->find($categoryId);
+        if (!$category) {
+            throw new NotFoundHttpException('Category not found');
+        }
+
+        $note = new Note($title, $text, $category);
+        $em->persist($note);
+        $em->flush();
+
+        $this->addFlash('success', sprintf('Note "%s" was created', $note->getTitle()));
+
+        return $this->redirectToRoute('notelist_create');
     }
 
     /**
-     * @Route("/delete_note/{id}", name="create")
+     * @Route("/delete/{id}", name="delete")
      */
-    public function deleteAction(int $id): Response
+    public function deleteAction(int $id, EntityManagerInterface $entityManager): Response
     {
-        $entityManager = $this->getDoctrine()->getManager();
         $noteToDelete = $this->noteRepository->find($id);
+        if (!$noteToDelete) {
+            throw new NotFoundHttpException('Note not found');
+        }
+
         $entityManager->remove($noteToDelete);
         $entityManager->flush();
 
-        return $this->render('notelist/delete_note.html.twig', [
-            'id' => $id,
-        ]);
+        $this->addFlash('success', sprintf('Note "%s" was deleted', $noteToDelete->getTitle()));
+
+        return $this->redirectToRoute('notelist_list_all');
     }
 }
